issue 1751: [Filestore] WriteBackCache refactoring - extract WriteBackCacheStats + update TPersistentStorageStats
