+++ cloud/filestore/libs/vfs_fuse/write_back_cache/write_back_cache.h
@@ -13,11 +13,17 @@
 
 namespace NCloud::NFileStore::NFuse {
 
+namespace NWriteBackCache {
+
 ////////////////////////////////////////////////////////////////////////////////
 
 struct IWriteBackCacheStats;
 using IWriteBackCacheStatsPtr =
     std::shared_ptr<IWriteBackCacheStats>;
+
+}   // namespace NWriteBackCache
+
+////////////////////////////////////////////////////////////////////////////////
 
 class TWriteBackCache final
 {
@@ -33,7 +39,7 @@ public:
         IFileStorePtr session,
         ISchedulerPtr scheduler,
         ITimerPtr timer,
-        IWriteBackCacheStatsPtr stats,
+        NWriteBackCache::IWriteBackCacheStatsPtr stats,
         TLog log,
         const TString& fileSystemId,
         const TString& clientId,
@@ -74,7 +80,6 @@ public:
     ui64 GetCachedNodeSize(ui64 nodeId) const;
     void SetCachedNodeSize(ui64 nodeId, ui64 size);
 
-    enum class EWriteDataRequestStatus;
     struct TPersistentQueueStats;
 
 private:
@@ -94,103 +99,5 @@ private:
     class TContiguousWriteDataEntryPartsReader;
     class TWriteDataEntryIntervalMap;
 };
-
-////////////////////////////////////////////////////////////////////////////////
-
-/**
- * WriteData request life cycle:
- * Initial -> Pending -> Cached -> (FlushRequested) -> Flushing -> Flushed
- *
- * Status FlushRequested may be skipped — Flush takes as much WriteData requests
- * as possible from |TWriteBackCache::TNodeState::CachedEntries|.
- *
- * For each NodeId it is guaranteed that there are no requests with out-of-order
- * statuses: if two requests A and B have the same NodeId, and the request A was
- * added to the queue later than B, then A.Status <= B.Status.
- */
-
-enum class TWriteBackCache::EWriteDataRequestStatus
-{
-    // The object has just been created and does not hold a request.
-    Initial,
-
-    // Restoration from the persisent buffer was failed.
-    // The request will not be processed further.
-    Corrupted,
-
-    // Write request is waiting for the conditions:
-    // - enough space in the persistent buffer to store the request;
-    // - no overlapping read requests in progress.
-    Pending,
-
-    // Write request has been stored in the persistent buffer
-    // The caller code observes the request as completed
-    Cached,
-
-    // Write request is being flushed
-    Flushing,
-
-    // Write request has been written to the session and can be removed from
-    // the persistent buffer
-    Flushed
-};
-
-////////////////////////////////////////////////////////////////////////////////
-
-struct TWriteBackCache::TPersistentQueueStats
-{
-    ui64 RawCapacity = 0;
-    ui64 RawUsedBytesCount = 0;
-    ui64 MaxAllocationBytesCount = 0;
-    bool IsCorrupted = false;
-};
-
-////////////////////////////////////////////////////////////////////////////////
-
-struct IWriteBackCacheStats
-{
-    enum class EReadDataRequestCacheStatus
-    {
-        // A request wasn't served from the cache
-        Miss,
-
-        // A request was partially served from the cache
-        PartialHit,
-
-        // A request was fully served from the cache
-        FullHit
-    };
-
-    virtual ~IWriteBackCacheStats() = default;
-
-    virtual void ResetNonDerivativeCounters() = 0;
-
-    virtual void FlushStarted() = 0;
-    virtual void FlushCompleted() = 0;
-    virtual void FlushFailed() = 0;
-
-    virtual void IncrementNodeCount() = 0;
-    virtual void DecrementNodeCount() = 0;
-
-    virtual void WriteDataRequestEnteredStatus(
-        TWriteBackCache::EWriteDataRequestStatus status) = 0;
-
-    virtual void WriteDataRequestExitedStatus(
-        TWriteBackCache::EWriteDataRequestStatus status,
-        TDuration duration) = 0;
-
-    virtual void WriteDataRequestUpdateMinTime(
-        TWriteBackCache::EWriteDataRequestStatus status,
-        TInstant minTime) = 0;
-
-    virtual void AddReadDataStats(
-        IWriteBackCacheStats::EReadDataRequestCacheStatus status,
-        TDuration pendingDuration) = 0;
-
-    virtual void UpdatePersistentQueueStats(
-        const TWriteBackCache::TPersistentQueueStats& stats) = 0;
-};
-
-IWriteBackCacheStatsPtr CreateDummyWriteBackCacheStats();
 
 }   // namespace NCloud::NFileStore::NFuse
