+++ cloud/filestore/libs/vfs_fuse/write_back_cache/write_back_cache_stats.cpp
@@ -1,0 +1,75 @@
+#include "write_back_cache_stats.h"
+
+namespace NCloud::NFileStore::NFuse::NWriteBackCache {
+
+namespace {
+
+////////////////////////////////////////////////////////////////////////////////
+
+class TDummyWriteBackCacheStats: public IWriteBackCacheStats
+{
+public:
+    void ResetNonDerivativeCounters() override
+    {}
+
+    void FlushStarted() override
+    {}
+
+    void FlushCompleted() override
+    {}
+
+    void FlushFailed() override
+    {}
+
+    void IncrementNodeCount() override
+    {}
+
+    void DecrementNodeCount() override
+    {}
+
+    void WriteDataRequestEnteredStatus(EWriteDataRequestStatus status) override
+    {
+        Y_UNUSED(status);
+    }
+
+    void WriteDataRequestExitedStatus(
+        EWriteDataRequestStatus status,
+        TDuration duration) override
+    {
+        Y_UNUSED(status);
+        Y_UNUSED(duration);
+    }
+
+    void WriteDataRequestUpdateMinTime(
+        EWriteDataRequestStatus status,
+        TInstant minTime) override
+    {
+        Y_UNUSED(status);
+        Y_UNUSED(minTime);
+    }
+
+    void AddReadDataStats(
+        EReadDataRequestCacheStatus status,
+        TDuration duration) override
+    {
+        Y_UNUSED(status);
+        Y_UNUSED(duration);
+    }
+
+    void UpdatePersistentStorageStats(
+        const TPersistentStorageStats& stats) override
+    {
+        Y_UNUSED(stats);
+    }
+};
+
+}   // namespace
+
+////////////////////////////////////////////////////////////////////////////////
+
+IWriteBackCacheStatsPtr CreateDummyWriteBackCacheStats()
+{
+    return std::make_shared<TDummyWriteBackCacheStats>();
+}
+
+}   // namespace NCloud::NFileStore::NFuse::NWriteBackCache
