+++ cloud/filestore/libs/vfs_fuse/write_back_cache/write_back_cache_util_ut.cpp
@@ -898,7 +898,7 @@ Y_UNIT_TEST_SUITE(TCalculateDataPartsToReadTest)
         constexpr ui64 MaxLength = 10;
         constexpr size_t MaxIntervalCount = 10;
 
-        auto stats = CreateDummyWriteBackCacheStats();
+        auto stats = NWriteBackCache::CreateDummyWriteBackCacheStats();
 
         for (size_t iter = 0; iter < IterationCount; iter++) {
             size_t intervalCount = RandomNumber(MaxIntervalCount) + 1;
